Change (#248): update name of dataset-resource
Renamed 'data_set' and 'data-set' to 'dataset'
to make the program consistent. Also renamed
files and folders.
