+++ deploy/ansible/hanami/roles/hanami/templates/hanami.conf.j2
@@ -5,7 +5,7 @@ database = "/etc/hanami/hanami_db"
 socket_path = "/tmp/hanami"
 
 [storage] 
-data_set_location = {{data_set_location}} 
+dataset_location = {{dataset_location}} 
 checkpoint_location = {{checkpoint_location}}
 tempfile_location = {{tempfile_location}}
 tempfile_timeout = 10
