+++ src/Hanami/src/api/http/v1/data_files/mnist/create_mnist_dataset.cpp
@@ -1,5 +1,5 @@
 /**
- * @file        create_mnist_data_set.cpp
+ * @file        create_mnist_dataset.cpp
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
@@ -20,16 +20,16 @@
  *      limitations under the License.
  */
 
-#include "create_mnist_data_set.h"
+#include "create_mnist_dataset.h"
 
 #include <core/temp_file_handler.h>
-#include <database/data_set_table.h>
+#include <database/dataset_table.h>
 #include <hanami_common/files/binary_file.h>
 #include <hanami_config/config_handler.h>
 #include <hanami_crypto/common.h>
 #include <hanami_root.h>
 
-CreateMnistDataSet::CreateMnistDataSet() : Blossom("Init new mnist-file data-set.")
+CreateMnistDataSet::CreateMnistDataSet() : Blossom("Init new mnist-file dataset.")
 {
     //----------------------------------------------------------------------------------------------
     // input
@@ -52,18 +52,18 @@ CreateMnistDataSet::CreateMnistDataSet() : Blossom("Init new mnist-file data-set
     // output
     //----------------------------------------------------------------------------------------------
 
-    registerOutputField("uuid", SAKURA_STRING_TYPE).setComment("UUID of the new data-set.");
+    registerOutputField("uuid", SAKURA_STRING_TYPE).setComment("UUID of the new dataset.");
 
-    registerOutputField("name", SAKURA_STRING_TYPE).setComment("Name of the new data-set.");
+    registerOutputField("name", SAKURA_STRING_TYPE).setComment("Name of the new dataset.");
 
     registerOutputField("owner_id", SAKURA_STRING_TYPE)
-        .setComment("ID of the user, who created the data-set.");
+        .setComment("ID of the user, who created the dataset.");
 
     registerOutputField("project_id", SAKURA_STRING_TYPE)
-        .setComment("ID of the project, where the data-set belongs to.");
+        .setComment("ID of the project, where the dataset belongs to.");
 
     registerOutputField("visibility", SAKURA_STRING_TYPE)
-        .setComment("Visibility of the data-set (private, shared, public).");
+        .setComment("Visibility of the dataset (private, shared, public).");
 
     registerOutputField("type", SAKURA_STRING_TYPE).setComment("Type of the new set (mnist)");
 
@@ -92,7 +92,7 @@ CreateMnistDataSet::runTask(BlossomIO& blossomIO,
 
     // get directory to store data from config
     bool success = false;
-    std::string targetFilePath = GET_STRING_CONFIG("storage", "data_set_location", success);
+    std::string targetFilePath = GET_STRING_CONFIG("storage", "dataset_location", success);
     if (success == false) {
         status.statusCode = INTERNAL_SERVER_ERROR_RTYPE;
         error.addMeesage("file-location to store dataset is missing in the config");
