+++ src/Hanami/src/api/http/v1/data_files/mnist/finalize_mnist_dataset.cpp
@@ -1,5 +1,5 @@
 /**
- * @file        finalize_mnist_data_set.cpp
+ * @file        finalize_mnist_dataset.cpp
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
@@ -20,20 +20,20 @@
  *      limitations under the License.
  */
 
-#include "finalize_mnist_data_set.h"
+#include "finalize_mnist_dataset.h"
 
 #include <core/temp_file_handler.h>
-#include <database/data_set_table.h>
+#include <database/dataset_table.h>
 #include <hanami_common/files/binary_file.h>
 #include <hanami_common/methods/file_methods.h>
 #include <hanami_crypto/common.h>
-#include <hanami_files/data_set_files/data_set_file.h>
-#include <hanami_files/data_set_files/image_data_set_file.h>
+#include <hanami_files/dataset_files/dataset_file.h>
+#include <hanami_files/dataset_files/image_dataset_file.h>
 #include <hanami_root.h>
 
 FinalizeMnistDataSet::FinalizeMnistDataSet()
     : Blossom(
-        "Finalize uploaded data-set by checking completeness of the "
+        "Finalize uploaded dataset by checking completeness of the "
         "uploaded and convert into generic format.")
 {
     errorCodes.push_back(NOT_FOUND_RTYPE);
@@ -43,7 +43,7 @@ FinalizeMnistDataSet::FinalizeMnistDataSet()
     //----------------------------------------------------------------------------------------------
 
     registerInputField("uuid", SAKURA_STRING_TYPE)
-        .setComment("UUID of the new data-set.")
+        .setComment("UUID of the new dataset.")
         .setRegex(UUID_REGEX);
 
     registerInputField("uuid_input_file", SAKURA_STRING_TYPE)
@@ -58,7 +58,7 @@ FinalizeMnistDataSet::FinalizeMnistDataSet()
     // output
     //----------------------------------------------------------------------------------------------
 
-    registerOutputField("uuid", SAKURA_STRING_TYPE).setComment("UUID of the new data-set.");
+    registerOutputField("uuid", SAKURA_STRING_TYPE).setComment("UUID of the new dataset.");
 
     //----------------------------------------------------------------------------------------------
     //
@@ -134,7 +134,7 @@ FinalizeMnistDataSet::runTask(BlossomIO& blossomIO,
  * @brief convert mnist-data into generic format
  *
  * @param filePath path to the resulting file
- * @param name data-set name
+ * @param name dataset name
  * @param inputBuffer buffer with input-data
  * @param labelBuffer buffer with label-data
  *
