+++ src/frontend/Hanami-Dashboard/src/subsites/kyouko/task.html
@@ -241,7 +241,7 @@ limitations under the License. -->
 
     taskTypes = ["train", "request"];
 
-    fillDropdownList("dropdown_div", "/control/v1/data_set/all");
+    fillDropdownList("dropdown_div", "/control/v1/dataset/all");
     fillStaticDropdownList("dropdown_taskType_div", taskTypes);
