+++ src/frontend/Hanami-Dashboard/src/subsites/shiori/dataset.html
@@ -154,7 +154,7 @@ limitations under the License. -->
 <script src="/scripts/table_processing.js"></script>
 <script src="/scripts/common.js"></script>
 <script src="/scripts/actions.js"></script>
-<script src="/sdk/javascript/data_set.js"></script> 
+<script src="/sdk/javascript/dataset.js"></script> 
 <script>
     
     /**
