+++ src/libraries/hanami_files/include/hanami_files/dataset_files/dataset_file.h
@@ -1,5 +1,5 @@
 /**
- * @file        data_set_file.cpp
+ * @file        dataset_file.cpp
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
