+++ src/libraries/hanami_files/include/hanami_files/dataset_files/image_dataset_file.h
@@ -1,5 +1,5 @@
 /**
- * @file        table_data_set_file.cpp
+ * @file        image_dataset_file.h
  *
  * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
  *
@@ -20,54 +20,34 @@
  *      limitations under the License.
  */
 
-#ifndef HANAMI_TABLEDATASETFILE_H
-#define HANAMI_TABLEDATASETFILE_H
+#ifndef HANAMI_IMAGEDATASETFILE_H
+#define HANAMI_IMAGEDATASETFILE_H
 
-#include <hanami_files/data_set_files/data_set_file.h>
+#include <hanami_files/dataset_files/dataset_file.h>
 
-class TableDataSetFile : public DataSetFile
+class ImageDataSetFile : public DataSetFile
 {
    public:
-    struct TableTypeHeader {
-        uint64_t numberOfColumns = 0;
-        uint64_t numberOfLines = 0;
+    struct ImageTypeHeader {
+        uint64_t numberOfInputsX = 0;
+        uint64_t numberOfInputsY = 0;
+        uint64_t numberOfOutputs = 0;
+        uint64_t numberOfImages = 0;
     };
 
-    struct TableHeaderEntry {
-        char name[256];
-        bool isInput = false;
-        bool isOutput = false;
-        float multiplicator = 1.0f;
-        float averageVal = 0.0f;
-        float maxVal = 0.0f;
-
-        void setName(const std::string& name)
-        {
-            uint32_t nameSize = name.size();
-            if (nameSize > 255) {
-                nameSize = 255;
-            }
-            memcpy(this->name, name.c_str(), nameSize);
-            this->name[nameSize] = '\0';
-        }
-    };
-
-    TableDataSetFile(const std::string& filePath);
-    TableDataSetFile(Hanami::BinaryFile* file);
-    ~TableDataSetFile();
+    ImageDataSetFile(const std::string& filePath);
+    ImageDataSetFile(Hanami::BinaryFile* file);
+    ~ImageDataSetFile();
     bool updateHeader(Hanami::ErrorContainer& error);
     bool getPayload(Hanami::DataBuffer& result,
                     Hanami::ErrorContainer& error,
                     const std::string& columnName = "");
 
-    void print();
-
-    TableTypeHeader tableHeader;
-    std::vector<TableHeaderEntry> tableColumns;
+    ImageTypeHeader imageHeader;
 
    protected:
     void initHeader();
     void readHeader(const uint8_t* u8buffer);
 };
 
-#endif  // HANAMI_TABLEDATASETFILE_H
+#endif  // HANAMI_IMAGEDATASETFILE_H
